+++ src/Console.php
@@ -82,7 +82,11 @@ class Console implements ContainerAwareInterface
 						throw new \RuntimeException(sprintf('Required class "%s" not found.', $className));
 					}
 
-					$commands[strtolower("$namespace:" . str_replace('Command', '', $command))] = $this->getContainer()->get($className);
+					// If the class isn't instantiable, it isn't a valid command
+					if ((new \ReflectionClass($className))->isInstantiable())
+					{
+						$commands[strtolower("$namespace:" . str_replace('Command', '', $command))] = $this->getContainer()->get($className);
+					}
 				}
 			}
 			else
@@ -95,7 +99,11 @@ class Console implements ContainerAwareInterface
 					throw new \RuntimeException(sprintf('Required class "%s" not found.', $className));
 				}
 
-				$commands[strtolower(str_replace('Command', '', $command))] = $this->getContainer()->get($className);
+				// If the class isn't instantiable, it isn't a valid command
+				if ((new \ReflectionClass($className))->isInstantiable())
+				{
+					$commands[strtolower(str_replace('Command', '', $command))] = $this->getContainer()->get($className);
+				}
 			}
 		}
